src/main/java/ua/nure/ipz/zoo/entity/Ration.java
package ua.nure.ipz.zoo.entity;

import ua.nure.ipz.zoo.util.DomainEntity;

import javax.persistence.CascadeType;
import javax.persistence.Entity;
import javax.persistence.OneToMany;
import javax.persistence.OneToOne;
import java.util.ArrayList;
import java.util.List;

@Entity
public class Ration extends DomainEntity {
    @OneToOne(cascade = CascadeType.PERSIST)
    private Animal animal;
    @OneToMany(cascade = CascadeType.PERSIST)
    private List<Food> foods = new ArrayList<>();

    public Ration() {
    }

    public Ration(Animal animal) {
        this.animal = animal;
    }

    public Animal getAnimal() {
        return animal;
    }

    public void setAnimal(Animal animal) {
        this.animal = animal;
    }

    public List<Food> getFoods() {
        return foods;
    }

    public void setFoods(List<Food> foods) {
        this.foods = foods;
    }
}
